chore(api documentation) Add swagger for documentation
