+++ server/lib/upload.js
@@ -1,0 +1,26 @@
+import multer from 'multer';
+
+const storage = multer.diskStorage({
+  filename: (req, file, cb) => {
+    cb(null, Date.now() + file.originalname);
+  },
+});
+const fileFilter = (req, file, cb) => {
+  if (
+    file.mimetype === 'image/jpg'
+    || file.mimetype === 'image/png'
+    || file.mimetype === 'image/jpeg'
+  ) {
+    cb(null, true);
+  } else {
+    cb(null, false);
+  }
+};
+
+const upload = multer({
+  storage,
+  fileFilter,
+  limits: { fileSize: 500000 },
+});
+
+export default upload;
