+++ swagger.json
@@ -1,0 +1,17 @@
+{
+  "openapi": "3.0.0",
+  "x-api-id": "automart",
+  "info": {
+    "version": "1.0.0",
+    "title": "automart",
+    "license": {
+      "name": "ISC"
+    }
+  },
+  "servers": [
+    {
+      "url": "https://github.com/tvpeter/automart"
+    }
+  ],
+  "paths": {}
+}
